Update requests:
The require a custom table size option needs to be at the top of all the variants as it needs to be closest to the native size dropdowns.

The current top section that’s been called Edge Profile should be called Timber Type

On the choose edge profile section change Wooden Planed to Modern Planed.

On the thickness section can the 50mm box be formatted to be the same sized box as the 35mm?

Can the default colour that is chosen upon entry be Forest brown please, as its that finish that has the most distinctive variance in the rustic and modern planed section.

Can the warning text box that says “for tables over 180cm we recommend a thickness of 50mm! be reformatted to a light grey back ground with the exclamation mark background being black instead of red?

Not included:
The hard wax oil (£40) live edge (£30) custom RAL (£60) and spider legs (£200) choices are all priced variants, I assume its possible for varints to be priced and for those prices to be added to the shopify native price during the product design process?

The table image that appears when raw steel and black is chosen isn’t required. Id rather just the description of each leg finish pop up when the leg finishes are picked

The Custom RAL colour appears as “Unwelded” when its pressed, please change that to Custom RAL

Please add the correct description for the custom RAL section, see attached screen shot.

I think we should be ready to go live with the above changes.